[feat]: 공고상세페이지 분기
제거 여러번 해매서 아직 하드코딩 한 부분은 못바꿔서 마이페이지 먼저
해볼려구요..
