+++ src/styles/DeleteSpin.css
@@ -1,0 +1,6 @@
+/* input 스핀버튼 없애는거 */
+input[type='number']::-webkit-outer-spin-button,
+input[type='number']::-webkit-inner-spin-button {
+  -webkit-appearance: none;
+  margin: 0;
+}
